segundo commit alumnat

+++ src/cat/iesjoaquimmir/alumnat/views/model/businesslayer/utilities/NumberUtils.java
@@ -2,6 +2,6 @@
 package cat.iesjoaquimmir.alumnat.views.model.businesslayer.utilities;
 
 public class NumberUtils {
-    public static final long UNSAVED_VALUE = -1;
+    public static final int UNSAVED_VALUE = -1;
     public static final int ZERO_VALUE = 0;
 }
